Refactor module book files
