+++ frontend/package.json
@@ -7,6 +7,7 @@
     "dev": "vite --host",
     "debug:deploy": "echo '=> files and dirs:' && find . \\( -path ./node_modules -o -path ./dist \\) -prune -o -type d -o -type f -print",
     "build": "npm run build:css && tsc -b && vite build",
+    "build:test": "vite build",
     "build:prod": "npm run build:css && tsc -p tsconfig.json && vite build --mode production",
     "build:css": "npm run build:css-light && npm run build:css-dark",
     "build:css-light": "npx @tailwindcss/cli -i ./src/common/styles/main.light.css -o ./public/main.light.css --mode=production",
@@ -25,7 +26,7 @@
     "@headlessui/react": "^2.2.0",
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^3.10.0",
-    "@radix-ui/react-hover-card": "^1.1.4",
+    "@radix-ui/react-hover-card": "^1.1.6",
     "@radix-ui/themes": "^3.1.6",
     "@reduxjs/toolkit": "^2.5.1",
     "@tailwindcss/cli": "^4.0.4",
