+++ frontend/src/common/root.reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import bookReducer from '@/module/book/book.reducer';
-import graphicAssetsReducer from '@/module/book/sidePanel/graphicAssets/graphicAssets.reducer';
+import bookReducer from '@/module/book/Book.slice';
+import graphicAssetsReducer from '@/module/book/sidePanel/graphicAssets/GraphicAssets.slice';
 
 const rootReducer = combineReducers({
   book: bookReducer,
