+++ frontend/src/common/styles/main.base.css
@@ -6,7 +6,6 @@
   --container-4xs: 12rem;
 }
 
-
 /* —————————————————————————————————————————————————————————————————————————————
   radix-ui configuration
   ———————————————————————————————————————————————————————————————————————————— */
@@ -39,70 +38,22 @@ body {
   overflow: hidden;
 }
 
-@layer base {
-  div[data-is-root-theme].dark {
-    background: radial-gradient(
-      circle at top left,
-      theme('colors.secondary.800') 2%,
-      theme('colors.primary.950') 50%
-    );
-  }
-  div[data-is-root-theme].light {
-    background: radial-gradient(
-      circle at top left,
-      theme('colors.secondary.200') 2%,
-      theme('colors.primary.50') 50%
-    );
-  }
+div[data-is-root-theme].dark {
+  background: radial-gradient(
+    circle at top left,
+    var(--color-secondary-800) 2%,
+    var(--color-primary-950) 50%
+  );
+}
+div[data-is-root-theme].light {
+  background: radial-gradient(
+    circle at top left,
+    var(--color-secondary-200) 2%,
+    var(--color-primary-50) 50%
+  );
 }
 
 #root {
   min-height: 100vh;
   min-width: 100vw;
 }
-
-/* —————————————————————————————————————————————————————————————————————————————
-  components
-  ———————————————————————————————————————————————————————————————————————————— */
-
-/* —————————————————————————————————————————————————————————————————————————————
-  Transitions
-  ———————————————————————————————————————————————————————————————————————————— */
-/* ::view-transition-old(root) {
-  animation: none;
-}
-
-::view-transition-new(root) {
-  animation: none;
-}
-
-::view-transition-old(*),
-::view-transition-new(*) {
-  animation-duration: 0.5s;
-}
-
-@keyframes fade-in {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes fade-out {
-  from {
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-  }
-}
-
-::view-transition-old(root) {
-  animation: fade-out 0.5s;
-}
-
-::view-transition-new(root) {
-  animation: fade-in 0.5s;
-} */
