+++ frontend/src/common/styles/main.light.css
@@ -1,4 +1,30 @@
 @import 'tailwindcss';
+@theme {
+  --color-primary-50: #f5f5f9;
+  --color-primary-100: #e7e8f2;
+  --color-primary-200: #d5d6e8;
+  --color-primary-300: #b9bbd7;
+  --color-primary-400: #979ac3;
+  --color-primary-500: #7e7db4;
+  --color-primary-600: #706ba5;
+  --color-primary-700: #645d92;
+  --color-primary-800: #58517c;
+  --color-primary-900: #494464;
+  --color-primary-950: #302d3e;
+
+  --color-secondary-50: #fff1f2;
+  --color-secondary-100: #ffe4e6;
+  --color-secondary-200: #fecdd3;
+  --color-secondary-300: #fda4af;
+  --color-secondary-400: #fb7186;
+  --color-secondary-500: #f43f5f;
+  --color-secondary-600: #e11d49;
+  --color-secondary-700: #be123d;
+  --color-secondary-800: #9f123a;
+  --color-secondary-900: #881338;
+  --color-secondary-950: #4c0519;
+}
+
 @config "./../../../tailwind.config.js";
 @import './fonts.css';
 @import './main.base.css';
