+++ frontend/src/module/book/canvas/element/types/Circle.ts
@@ -34,6 +34,7 @@ export class Circle implements DrawableElement {
   }
 
   draw(canvas: fabric.Canvas): fabric.Object {
+    console.log('draw', canvas);
     // canvas.add(this.circle);
     return this.circle;
   }
