--- frontend/src/module/book/components/SidePanel/SidePanel.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-export const SidePanel: React.FC<{
-  className?: string;
-  children?: React.ReactNode;
-}> = ({ className, children }) => {
-  return <aside className={`${className || ''}`}>{children}</aside>;
-};
