+++ frontend/src/module/book/sidePanel/SidePanel.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import {
   PaintBrushIcon,
   RectangleGroupIcon,
@@ -6,10 +6,10 @@ import {
 } from '@heroicons/react/24/outline';
 
 import TextIcon from '@assets/icons/icon_text.svg?react';
-import { TabType, Tab } from '../sidepanel.types';
+import { TabType, Tab } from './Sidepanel.types';
 import Tabs from './Tabs';
-import UserContent from '../user/UserContent';
-import ElementContent from '../element/ElementContent';
+import UserContent from './user/UserContent';
+import ElementContent from './element/ElementContent';
 import useLocalStorage from '@/common/hooks/useLocalStorage';
 
 const tabs: Tab[] = [
