+++ frontend/src/module/book/sidePanel/sidepanel.state.ts
@@ -1,5 +1,4 @@
-// import { RootState } from '@/common/store';
-import { TabType } from './sidepanel.types';
+import { TabType } from './Sidepanel.types';
 interface SidePanelState {
   isOpen: true;
   width: number; // In pixel
@@ -10,14 +9,6 @@ const initialState: SidePanelState = {
   isOpen: true,
   width: 500,
   selectedTab: TabType.Element,
-  // areLocalUpdatesSaved: true,
-  // refreshGraphics: false,
-  // graphicAssets: [],
-  // isLoading: false,
-  // error: null,
 };
-
-// export const selectIsLoading = (state: RootState) => state.book.isLoading;
-// export const selectBook = (state: RootState) => state.book;
 
 export default initialState;
